comment responses and test fixes
