+++ weld/llvm.rs
@@ -24,14 +24,6 @@ use super::util::IdGenerator;
 
 #[cfg(test)]
 use super::parser::*;
-#[cfg(test)]
-use super::weld_run_free;
-#[cfg(test)]
-use super::weld_rt_malloc;
-#[cfg(test)]
-use super::weld_rt_realloc;
-#[cfg(test)]
-use super::weld_rt_free;
 
 static PRELUDE_CODE: &'static str = include_str!("resources/prelude.ll");
 static VECTOR_CODE: &'static str = include_str!("resources/vector.ll");
@@ -439,14 +431,18 @@ impl LlvmGenerator {
                 let bound_cmp = wrap_ctx.var_ids.next();
                 let mut grain_size = 1024;
                 if par_for.innermost {
-                    wrap_ctx.code.add(format!("{} = icmp ule i64 {}, {}", bound_cmp, upper_bound,
-                        grain_size));
-                    wrap_ctx.code.add(format!("br i1 {}, label %for.ser, label %for.par", bound_cmp));
+                    wrap_ctx.code.add(format!("{} = icmp ule i64 {}, {}",
+                                              bound_cmp,
+                                              upper_bound,
+                                              grain_size));
+                    wrap_ctx.code
+                        .add(format!("br i1 {}, label %for.ser, label %for.par", bound_cmp));
                     wrap_ctx.code.add(format!("for.ser:"));
                     let mut body_arg_types = try!(self.get_arg_str(&func.params, ""));
                     body_arg_types.push_str(format!(", i64 0, i64 {}", upper_bound).as_str());
                     wrap_ctx.code.add(format!("call void @f{}({})", func.id, body_arg_types));
-                    let cont_arg_types = try!(self.get_arg_str(&sir.funcs[par_for.cont].params, ""));
+                    let cont_arg_types =
+                        try!(self.get_arg_str(&sir.funcs[par_for.cont].params, ""));
                     wrap_ctx.code.add(format!("call void @f{}({})", par_for.cont, cont_arg_types));
                     wrap_ctx.code.add(format!("br label %fn.end"));
                 } else {
@@ -1512,825 +1508,4 @@ fn types() {
 
     let struct2 = parse_type("{i32,bool}").unwrap().to_type().unwrap();
     assert_eq!(gen.llvm_type(&struct2).unwrap(), "%s1");
-}
-
-#[test]
-fn runtime_functions() {
-    weld_rt_free(0, weld_rt_realloc(0, weld_rt_malloc(0, 16), 32));
-}
-
-#[test]
-fn basic_program() {
-    let code = "|| 40 + 2";
-
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let inp = Box::new(WeldInputArgs {
-        input: 0,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 42);
-    weld_run_free(-1);
-}
-
-#[test]
-fn f64_cast() {
-    let code = "|| f64(40 + 2)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-
-    let inp = Box::new(WeldInputArgs {
-        input: 0,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result = module.run(ptr) as *const f64;
-    let result = unsafe { *result };
-    assert_eq!(result, 42.0);
-    weld_run_free(-1);
-}
-
-#[test]
-fn i32_cast() {
-    let code = "|| i32(0.251 * 4.0)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let inp = Box::new(WeldInputArgs {
-        input: 0,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 1);
-    weld_run_free(-1);
-}
-
-#[test]
-fn program_with_args() {
-    let code = "|x:i32| 40 + x";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input: i32 = 2;
-    let inp = Box::new(WeldInputArgs {
-        input: &input as *const i32 as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 42);
-    weld_run_free(-1);
-}
-
-#[test]
-fn let_statement() {
-    let code = "|x:i32| let y = 40 + x; y + 2";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input: i32 = 2;
-    let inp = Box::new(WeldInputArgs {
-        input: &input as *const i32 as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 44);
-    weld_run_free(-1);
-}
-
-#[test]
-fn if_statement() {
-    let code = "|x:i32| if(true, 3, 4)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input: i32 = 2;
-    let inp = Box::new(WeldInputArgs {
-        input: &input as *const i32 as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-
-    let ptr = Box::into_raw(inp) as i64;
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 3);
-    weld_run_free(-1);
-}
-
-#[test]
-fn comparison() {
-    let code = "|x:i32| if(x>10, x, 10)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let mut input: i32 = 2;
-    let inp = Box::new(WeldInputArgs {
-        input: &input as *const i32 as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 10);
-    weld_run_free(-1);
-
-    input = 20;
-    let result = module.run(ptr) as *const i32;
-    let result = unsafe { *result };
-    assert_eq!(result, 20);
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_vector_lookup() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-    }
-
-    let code = "|x:vec[i32]| lookup(x, 3L)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input = [1, 2, 4, 5];
-    let args = Args {
-        x: Vec {
-            data: &input as *const i32,
-            len: 4,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    let output = input[3];
-    assert_eq!(output, result);
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_for_appender_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        a: i32,
-    }
-
-    let code = "|x:vec[i32], a:i32| let b=a+1; map(x, |e| e+b)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input = [1, 2];
-    let args = Args {
-        x: Vec {
-            data: &input as *const i32,
-            len: 2,
-        },
-        a: 1,
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const Vec;
-    let result = unsafe { (*result_raw).clone() };
-    let output = [3, 4];
-    for i in 0..(result.len as isize) {
-        assert_eq!(unsafe { *result.data.offset(i) }, output[i as usize])
-    }
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_parallel_for_appender_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct WeldVec {
-        data: *const i32,
-        len: i64,
-    }
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct WeldVec64 {
-        data: *const i64,
-        len: i64,
-    }
-    let code = "|x:vec[i32]| result(for(x, appender[i64], |b,i,e| merge(b, i)))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let size: i32 = 10000;
-    let input: Vec<i32> = vec![0; size as usize];
-    let args = WeldVec {
-        data: input.as_ptr() as *const i32,
-        len: size as i64,
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const WeldVec as i64,
-        nworkers: 4,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const WeldVec64;
-    let result = unsafe { (*result_raw).clone() };
-    assert_eq!(result.len as i32, size);
-    for i in 0..(result.len as isize) {
-        assert_eq!(unsafe { *result.data.offset(i) }, i as i64)
-    }
-    weld_run_free(-1);
-}
-
-#[test]
-fn complex_parallel_for_appender_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct WeldVec {
-        data: *const i32,
-        len: i64,
-    }
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct WeldVec64 {
-        data: *const i64,
-        len: i64,
-    }
-    let code = "|x:vec[i32]| let a=appender[i64]; let b=merge(a,0L); let r=for(x,b,|b,i,e| \
-                let c=merge(b,1L); let d=for(x,c,|b,i,e| if(i<1L, merge(b,i), b)); merge(d, 2L)); \
-                result(merge(r,3L))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let size: i32 = 3000;
-    let input: Vec<i32> = vec![0; size as usize];
-    let args = WeldVec {
-        data: input.as_ptr() as *const i32,
-        len: size as i64,
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const WeldVec as i64,
-        nworkers: 4,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const WeldVec64;
-    let result = unsafe { (*result_raw).clone() };
-    assert_eq!(result.len as i32, size * 3 + 2);
-    assert_eq!(unsafe { *result.data.offset(0) }, 0);
-    assert_eq!(unsafe { *result.data.offset((size * 3 + 1) as isize) }, 3);
-    for i in 0..(size as isize) {
-        assert_eq!(unsafe { *result.data.offset(i * 3 + 1) }, 1);
-        assert_eq!(unsafe { *result.data.offset(i * 3 + 2) }, 0);
-        assert_eq!(unsafe { *result.data.offset(i * 3 + 3) }, 2)
-    }
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_for_merger_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        a: i32,
-    }
-
-    let code = "|x:vec[i32], a:i32| result(for(x, merger[i32,+], |b,i,e| merge(b, e+a)))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input = [1, 2, 3, 4, 5];
-    let args = Args {
-        x: Vec {
-            data: &input as *const i32,
-            len: 5,
-        },
-        a: 1,
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    let output = 20;
-    assert_eq!(result, output);
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_for_dictmerger_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Pair {
-        ele1: i32,
-        ele2: i32,
-    }
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct VecPrime {
-        data: *const Pair,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        y: Vec,
-    }
-
-    let code = "|x:vec[i32], y:vec[i32]| tovec(result(for(zip(x,y), dictmerger[i32,i32,+], \
-                |b,i,e| merge(b, e))))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let keys = [1, 2, 2, 1, 3];
-    let values = [2, 3, 4, 2, 1];
-    let args = Args {
-        x: Vec {
-            data: &keys as *const i32,
-            len: 5,
-        },
-        y: Vec {
-            data: &values as *const i32,
-            len: 5,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const VecPrime;
-    let result = unsafe { (*result_raw).clone() };
-    let output_keys = [1, 2, 3];
-    let output_values = [4, 7, 1];
-    for i in 0..(output_keys.len() as isize) {
-        let mut success = false;
-        let key = unsafe { (*result.data.offset(i)).ele1 };
-        let value = unsafe { (*result.data.offset(i)).ele2 };
-        for j in 0..(output_keys.len()) {
-            if output_keys[j] == key {
-                if output_values[j] == value {
-                    success = true;
-                }
-            }
-        }
-        assert_eq!(success, true);
-    }
-    assert_eq!(result.len, output_keys.len() as i64);
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_parallel_for_dictmerger_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Pair {
-        ele1: i32,
-        ele2: i32,
-    }
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct VecPrime {
-        data: *const Pair,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        y: Vec,
-    }
-
-    let code = "|x:vec[i32], y:vec[i32]| tovec(result(for(zip(x,y), dictmerger[i32,i32,+], \
-                |b,i,e| merge(b, e))))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    const DICT_SIZE: usize = 8192;
-    let mut keys = [0; DICT_SIZE];
-    let mut values = [0; DICT_SIZE];
-    for i in 0..DICT_SIZE {
-        keys[i] = i as i32;
-        values[i] = i as i32;
-    }
-    let args = Args {
-        x: Vec {
-            data: &keys as *const i32,
-            len: DICT_SIZE as i64,
-        },
-        y: Vec {
-            data: &values as *const i32,
-            len: DICT_SIZE as i64,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 4,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const VecPrime;
-    let result = unsafe { (*result_raw).clone() };
-    let output_keys = keys;
-    let output_values = values;
-    for i in 0..(output_keys.len() as isize) {
-        let mut success = false;
-        let key = unsafe { (*result.data.offset(i)).ele1 };
-        let value = unsafe { (*result.data.offset(i)).ele2 };
-        for j in 0..(output_keys.len()) {
-            if output_keys[j] == key {
-                if output_values[j] == value {
-                    success = true;
-                }
-            }
-        }
-        assert_eq!(success, true);
-    }
-    assert_eq!(result.len, output_keys.len() as i64);
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_dict_lookup() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        y: Vec,
-    }
-
-    let code = "|x:vec[i32], y:vec[i32]| let a = result(for(zip(x,y), dictmerger[i32,i32,+], \
-                |b,i,e| merge(b, e))); lookup(a, 1)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let keys = [1, 2, 2, 1, 3];
-    let values = [2, 3, 4, 2, 1];
-    let args = Args {
-        x: Vec {
-            data: &keys as *const i32,
-            len: 5,
-        },
-        y: Vec {
-            data: &values as *const i32,
-            len: 5,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    let output = 4;
-    assert_eq!(output, result);
-    weld_run_free(-1);
-}
-
-#[test]
-fn simple_length() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-    }
-
-    let code = "|x:vec[i32]| len(x)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let data = [2, 3, 4, 2, 1];
-    let args = Args {
-        x: Vec {
-            data: &data as *const i32,
-            len: 5,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    let output = 5;
-    assert_eq!(output, result);
-    weld_run_free(-1);
-}
-
-#[test]
-fn filter_length() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-    }
-
-    let code = "|x:vec[i32]| len(filter(x, |i| i < 4 && i > 1))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let data = [2, 3, 4, 2, 1];
-    let args = Args {
-        x: Vec {
-            data: &data as *const i32,
-            len: 5,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    let output = 3;
-    assert_eq!(output, result);
-    weld_run_free(-1);
-}
-
-#[test]
-fn flat_map_length() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-    }
-
-    let code = "|x:vec[i32]| len(flatten(map(x, |i:i32| x)))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-
-    let data = [2, 3, 4, 2, 1];
-    let args = Args {
-        x: Vec {
-            data: &data as *const i32,
-            len: 5,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    let output = 25;
-    assert_eq!(output, result);
-    weld_run_free(-1);
-}
-
-#[test]
-fn if_for_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        a: i32,
-    }
-
-    let code = "|x:vec[i32], a:i32| if(a > 5, map(x, |e| e+1), map(x, |e| e+2))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let input = [1, 2];
-
-    let args = Args {
-        x: Vec {
-            data: &input as *const i32,
-            len: 2,
-        },
-        a: 1,
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const Vec;
-    let result = unsafe { (*result_raw).clone() };
-    let output = [3, 4];
-    for i in 0..(result.len as isize) {
-        assert_eq!(unsafe { *result.data.offset(i) }, output[i as usize])
-    }
-    weld_run_free(-1);
-
-    let args = Args {
-        x: Vec {
-            data: &input as *const i32,
-            len: 2,
-        },
-        a: 6,
-    };
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const Vec;
-    let result = unsafe { (*result_raw).clone() };
-    let output = [2, 3];
-    for i in 0..(result.len as isize) {
-        assert_eq!(unsafe { *result.data.offset(i) }, output[i as usize])
-    }
-    weld_run_free(-1);
-}
-
-#[test]
-fn map_zip_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        y: Vec,
-    }
-
-    let code = "|x:vec[i32], y:vec[i32]| map(zip(x,y), |e| e.$0 + e.$1)";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let x = [1, 2, 3, 4];
-    let y = [5, 6, 7, 8];
-    let args = Args {
-        x: Vec {
-            data: &x as *const i32,
-            len: 4,
-        },
-        y: Vec {
-            data: &y as *const i32,
-            len: 2,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const Vec;
-    let result = unsafe { (*result_raw).clone() };
-    let output = [6, 8, 10, 12];
-    for i in 0..(result.len as isize) {
-        assert_eq!(unsafe { *result.data.offset(i) }, output[i as usize])
-    }
-    weld_run_free(-1);
-}
-
-#[test]
-fn iters_for_loop() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct Vec {
-        data: *const i32,
-        len: i64,
-    }
-    #[allow(dead_code)]
-    struct Args {
-        x: Vec,
-        y: Vec,
-    }
-
-    let code = "|x:vec[i32], y:vec[i32]| result(for(zip(iter(x,0L,4L,2L), y), appender, |b,i,e| \
-                merge(b,e.$0+e.$1)))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let x = [1, 2, 3, 4];
-    let y = [5, 6];
-    let args = Args {
-        x: Vec {
-            data: &x as *const i32,
-            len: 4,
-        },
-        y: Vec {
-            data: &y as *const i32,
-            len: 2,
-        },
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const Args as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const Vec;
-    let result = unsafe { (*result_raw).clone() };
-    let output = [6, 9];
-    for i in 0..(result.len as isize) {
-        assert_eq!(unsafe { *result.data.offset(i) }, output[i as usize])
-    }
-    weld_run_free(-1);
-}
-
-#[test]
-fn serial_parlib_test() {
-    #[derive(Clone)]
-    #[allow(dead_code)]
-    struct WeldVec {
-        data: *const i32,
-        len: i64,
-    }
-    let code = "|x:vec[i32]| result(for(x, merger[i32,+], |b,i,e| merge(b, e)))";
-    let module = compile_program(&parse_program(code).unwrap()).unwrap();
-    let size: i32 = 10000;
-    let input: Vec<i32> = vec![1; size as usize];
-    let args = WeldVec {
-        data: input.as_ptr() as *const i32,
-        len: size as i64,
-    };
-
-    let inp = Box::new(WeldInputArgs {
-        input: &args as *const WeldVec as i64,
-        nworkers: 1,
-        run_id: 0,
-    });
-    let ptr = Box::into_raw(inp) as i64;
-
-    let result_raw = module.run(ptr) as *const i32;
-    let result = unsafe { (*result_raw).clone() };
-    assert_eq!(result, size);
-    weld_run_free(-1);
 }
