+++ weld/resources/prelude.ll
@@ -18,17 +18,21 @@ declare void    @free_merger(i64, i8*)
 declare i32     @my_id_public()
 
 ; Parallel runtime structures
+; documentation in parlib.h
 %work_t = type { i8*, i64, i64, i64, i32, i64*, i64*, i32, i64, void (%work_t*)*, %work_t*, i32, i32, i32 }
+; documentation in vb.cpp
 %vb.vp = type { i8*, i64, i64, i64*, i64*, i32 }
 %vb.out = type { i8*, i64 }
 
 ; Input argument (input data pointer, nworkers, runId)
 %input_arg_t = type { i64, i32, i64 }
 
+; documenation in parlib.cpp
 declare void @set_result(i8*)
 declare i8* @get_result()
 declare void @pl_start_loop(%work_t*, i8*, i8*, void (%work_t*)*, void (%work_t*)*, i64, i64, i32)
 declare void @execute(void (%work_t*)*, i8*)
+; documentation in vb.cpp
 declare i8* @new_vb(i64, i64)
 declare void @new_piece(i8*, %work_t*)
 declare %vb.vp* @cur_piece(i8*, i32)
